+++ source/Structures/Spirits/Seasonal/AURORA/AURORA.ts
