+++ source/Structures/Spirits/Seasonal/AURORA/index.ts
@@ -1,7 +1,7 @@
-import AURORAGuide from "./AURORAGuide.js";
+import AURORA from "./AURORA.js";
 import MindfulMiner from "./MindfulMiner.js";
 import RunningWayfarer from "./RunningWayfarer.js";
 import SeedOfHope from "./SeedOfHope.js";
 import WarriorOfLove from "./WarriorOfLove.js";
 
-export default [AURORAGuide, RunningWayfarer, MindfulMiner, WarriorOfLove, SeedOfHope] as const;
+export default [AURORA, RunningWayfarer, MindfulMiner, WarriorOfLove, SeedOfHope] as const;
